prospeccao_leads/requirements.txt
aiohttp>=3.9.0
requests>=2.31.0
pandas>=2.1.0
gspread>=6.0.0
google-auth-oauthlib>=1.1.0
oauth2client>=4.1.3
tenacity>=8.2.0
python-dotenv>=1.0.0
